gpsd-machine-conf: add recipe
gpsd recipe in meta-oe is designed to have 'machine' specific dependencies
through gpsd-machine-conf recipe. It is a blank recipe by default, that can be
customized per machine.

In this recipe we customize it for dragonboard-410c and we pull in the right
dependencies needed for gpsd to work with the on board GPS, whose stack is
mostly running on the DSP.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>